% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTotalData.R
\name{buildTotalData}
\alias{buildTotalData}
\title{Function to build data frame with all information}
\usage{
buildTotalData(fbs_json, fcs_json)
}
\description{
Function to build data frame with all information
}
