% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScheduleData.R
\name{createScheduleData}
\alias{createScheduleData}
\title{Function to create data the table is built on}
\usage{
createScheduleData(
  total_schedule,
  game_wp_stats,
  game_odds_df,
  total_sp_plus,
  total_conf_info,
  rivalries_to_match
)
}
\arguments{
\item{total_schedule}{Data frame with schedule data from ESPN.com}

\item{game_wp_stats}{Data frame of game-level win probabilities from ESPN.com}

\item{game_odds_df}{Data frame of pre-game betting odds from ESPN.com}

\item{total_sp_plus}{Data frame of all SP+ rankings}

\item{total_conf_info}{Data frame of conference abbreviations, partially from `cfbFastR` and partially manually scraped}

\item{rivalries_to_match}{Data frame of rivalries scraped from wikipedia}
}
\description{
Function to create data the table is built on
}
