% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScheduleTable.R
\name{createScheduleTable}
\alias{createScheduleTable}
\title{Function to create display table for games}
\usage{
createScheduleTable(df, table_group_vars)
}
\arguments{
\item{df}{Dataframe with game data to be displayed}
}
\value{
a neat `gt()`
}
\description{
Function to create display table for games
}
