% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSPData.R
\name{createSPData}
\alias{createSPData}
\title{Function to clean SP+ data to be easily joined to ESPN games. Run once a week, then save to data}
\usage{
createSPData()
}
\description{
Function to clean SP+ data to be easily joined to ESPN games. Run once a week, then save to data
}
