% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbs_sp_plus.R
\docType{data}
\name{fbs_sp_plus}
\alias{fbs_sp_plus}
\title{SP+ scores from Bill Connelly.}
\format{
A data frame with 130 rows and 5 variables:
\describe{
  \item{TEAM}{Team name}
  \item{sp_plus}{Team-level SP+ values}
  \item{off_sp}{Offense SP+ values}
  \item{def_sp}{Defense SP+ values}
  \item{st_sp}{Special Teams SP+ values}
}
}
\source{
\url{https://www.espn.com/college-football/insider/story/_/id/30364909/college-football-sp+-rankings-week-12-bama-top-interesting-moves-big-ten}
}
\usage{
fbs_sp_plus
}
\description{
SP+ scores from Bill Connelly.
}
\keyword{datasets}
