% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGameInformation.R
\name{getGameInformation}
\alias{getGameInformation}
\title{Function to retrieve data for a particular game from ESPN.com}
\usage{
getGameInformation(game_id)
}
\arguments{
\item{game_id}{ID of game to be retrieved. Long integer, not entirely sure how they come up with these.}
}
\value{
One-row tibble with game_id along with dataframes with information on gambling odds, win probabilities, and box score information
}
\description{
Function to retrieve data for a particular game from ESPN.com
}
