% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBoxScore.R
\name{parseBoxScore}
\alias{parseBoxScore}
\title{Function to convert box-score JSON data from ESPN.com game page to named list}
\usage{
parseBoxScore(box_info)
}
\arguments{
\item{box_info}{JSON data from game page on ESPN.com}
}
\value{
Named list of statistics in box score
}
\description{
Function to convert box-score JSON data from ESPN.com game page to named list
}
