% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDivisionInfo.R
\name{parseDivisionInfo}
\alias{parseDivisionInfo}
\title{Convert ESPN.com JSON response to a useable dataframe}
\usage{
parseDivisionInfo(div_json)
}
\arguments{
\item{div_json}{Schedule for one NCAA FB division (FBS or FCS)}
}
\value{
Dataframe with one row of data per game
}
\description{
Convert ESPN.com JSON response to a useable dataframe
}
