% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGameInfo.R
\name{parseGameInfo}
\alias{parseGameInfo}
\title{Convert JSON of game info from ESPN.com to a tibble}
\usage{
parseGameInfo(game_json)
}
\arguments{
\item{game_json}{JSON data scraped from the ESPN.com schedule}
}
\value{
One-row tibble with information about game
}
\description{
Convert JSON of game info from ESPN.com to a tibble
}
