% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePlayData.R
\name{parsePlayData}
\alias{parsePlayData}
\title{Function to convert JSON data for one play to a tibble}
\usage{
parsePlayData(play_json)
}
\arguments{
\item{play_json}{JSON data of one play scraped from the game page on ESPN.com}
}
\value{
One-row tibble with information about that specific play
}
\description{
Function to convert JSON data for one play to a tibble
}
