% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePreGameOdds.R
\name{parsePreGameOdds}
\alias{parsePreGameOdds}
\title{Convert JSON of gambling odds to tibble}
\usage{
parsePreGameOdds(odds_json)
}
\arguments{
\item{odds_json}{JSON of gambling odds from ESPN.com game page}
}
\value{
One-row tibble with gambling odds from one provider
}
\description{
Convert JSON of gambling odds to tibble
}
